requirements.txt
PyQt6
chromadb
watchdog
PyMuPDF
pytesseract
Pillow
sentence-transformers
numpy
